Uniform device tree for differnt core configs
Instead of having multiple device tree files for differnt number of
cores, a single file is used. The device tree must be processes by a
C/C++ preprocessor before it is compiled. NRCPU must be passed to the
preprocessor. The compilation of the device trees is shifted from
buildroot to the postprocessing step.

Signed-off-by: Nils Bosbach <bosbach@ice.rwth-aachen.de>